rfac: Remove redundant home.dart and change related import statements
