+++ lib/login.dart
@@ -7,13 +7,14 @@ import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
+import 'Home.dart';
 import 'colors.dart';
 import 'help.dart';
 import 'package:appetizer/services/user.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:flutter_web_browser/flutter_web_browser.dart';
 
-import 'home.dart';
+import 'Home.dart';
 
 class Login extends StatefulWidget {
   final String code;
